[lti] Fix hint propagation with `super(...)` call
Summary:
This diff fixes the bug discovered in the diff before. We have `Decomp_CallSuper` for the `super(...)` and we originally do propagate it in statement.ml. However, this is lost when the source of truth regarding hints are moved to name_def.

Changelog: [fix] We will no longer complain about `missing-local-annot` when there is a proper typing context in `super` call. e.g. `super(x => x)` will now pass type checking under LTI if the super constructor call expects such function.

Reviewed By: panagosg7

Differential Revision: D43106743

fbshipit-source-id: a53d3e9640b599e743a91e5eb90a09729e3c527f