[IDE] Remove typechecking from document highlight command
Summary:
This diff pulls the typechecker completely out of the document highlight command and solely uses cached data from the typecheck that most of the IDE commands use. I had to pull out the data in the `obj_to_obj_map` to an essentially global level, even though it is only used by one command, so that it could be cached. This can be seen in the changes to `type_contents.ml`.

Changelog: Improve performance of the documentHighlight LSP command

Reviewed By: mroch

Differential Revision: D42825171

fbshipit-source-id: 781f1dbd9eed554ee2de8ecaf683a472946373a2