push post-reinit rechecks onto the recheck stream
Summary:
first let's define some terms: a "reinit" consists of a "reload" from saved state, followed by a "reparse" of the changes since the saved state was generated, followed by a "recheck" of the changes since `master`.

the reload, reparse and recheck are all in separate transactions.

if either the reparse or the recheck are canceled, the entire reinit is aborted, even though parts of it have been committed and can't be rolled back. the `env` is thrown away and we go back to the pre-reinit `env`.

`env.checked_files` in particular is now wrong: we think that files have been checked, because they were checked before the reinit, but they are no longer checked in shared memory. when we use `checked_files` to determine things like `unchanged_checked`, which is used to skip merge, we end up never checking files that we need to.

the fix in this diff is to schedule the reparse and recheck on the recheck stream, so that the "reload" transaction (almost) immediately ends the recheck after the transaction is committed. that ensures the new `env` takes over.

`recheck_loop` will then run the reparse and recheck (which probably got squashed into one recheck, yay) right away.

Changelog: [internal] this really only impacted reinits, since normal rechecks don't do a lot of work after commit

Reviewed By: jbrown215

Differential Revision: D43265619

fbshipit-source-id: 263c9a1b6e7b1e416999276ef3cb623c64ed464d