make recheck stream maintain physical equality
Summary:
when the recheck stream gets new events, they are merged into a single recheck workload. this workload will get run on the next time through the recheck loop.

when we get events we've seen before, like when multiple IDE commands have the same unchecked deps, we want to ignore it so that we don't trigger a cancelation just to restart an identical recheck.

we do this today via the `workload_changed` function, which inspects physical equality of each of the fields of the workload. there used to be other fields like profiling callbacks, which prevented us from using physical equality on the entire workload. those are gone, so now we can.

this is also necessary to set up the next diff

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D43265756

fbshipit-source-id: 68b15b1f491efff9994a3080ee21ce9a6dfe307a