fix changes-since computation in scm fetcher
Summary:
we run `process_updates` on all of the files that have changed since the saved state was created, because we need to recheck any of those files that we care about.

it expects absolute paths, but the scm fetcher was giving it relative paths, so it thought none of them were part of the flow root and so didn't care about any of them.

fix is to return absolute paths from the scm fetcher, like the other fetchers do.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D43326591

fbshipit-source-id: adb66c7be04f11fc85458bf6bf3484282672d790