[IDE] use cacheable obj_to_obj_hook in propertyFindRefs
Summary:
This diff removes the last typechecker hook from `propertyFindRefs` by replacing the `obj_to_obj_hook` with cached data from an earlier run of the typechecker hook. It replicates the behavior of the previous hook by building an inverse map of the obj_to_obj hook cached data (which is essentially what the old hook was doing).

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D42822256

fbshipit-source-id: 561bb174a26968fadc385dbb8a81bc532a3741cb