[IDE] Fix bug in document highlight with destructuring
Summary:
I noticed a bug in the DocumentHighlight command where it didn't handle highlighting on object keys that are renamed in a destructuring pattern.
```
let { x: _x, y: _y } = { x, y };
      ^ Hovering here does nothing
```
This is because the searcher (or typechecker hooks in the old version) don't consider destructuring patterns when looking for the object key that matches the target location. It works when the destructure is not renamed because it is found in the local var searcher.

This diff fixes the bug in the rewritten document highlight command. Since this is being rolled out next release, I'm not fixing the existing but in the implementation of document highlight using typechecker hooks.

Changelog: [Fix] DocumentHighlight now correctly highlights destructured object keys

Reviewed By: darichey

Differential Revision: D43202925

fbshipit-source-id: 6ad22890da7f089543fb3b4dfa86bcadb24cafc4