[IDE] Auto close jsx tags
Summary:
This is implemented with a non-standard LSP extension: a `flow/autoCloseJsx` request. When `>` is typed, the vscode extension will send this request with the position of the char. The flow LSP server can then search for an opening JSX tag that ends at that position and return a snippet for a corresponding closing tag. The vscode extension then inserts the snippet.

Changelog: [ide] Added option for automatically closing JSX tags.

Reviewed By: mroch

Differential Revision: D43244186

fbshipit-source-id: 8f2e6daad368be32fcb127c13fd775833e44bef9