[lti][perf] Optimize evaluation of Decomp_SentinelRefinement
Summary:
`Decomp_SentinelRefinement` is used to refine the hint when the object type is one branch of disjoint union. It should be doing nothing most of the time, but we are paying the cost of trying this refinement anyways.

In this diff, I optimized this by only doing refinement if the input type is some kind of union-like type. This is observed to cut the time spent on `Decomp_SentinelRefinement` by half.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D43337802

fbshipit-source-id: 62258516a119b2aa205a6af7346a7b3b15033233