Lazily map return hint for instantiation
Summary:
In the definition of hint decomposition ops, we have `Instantiate_Callee` and `Instantiate_Component`. They are supposed to do nothing and just behave like an identity function if the callee/component is not polymorphic. In terms of performance, what we care about is that they do no extra work at the moment we figure out that the callee/component in monomorphic.

This is mostly true, but not for return hint. We still eagerly convert a base hint of ast node to a base hint of type, which might have significant cost. This effort is completely wasted if the callee/component in monomorphic, but we will do the work anyways during env_resolution.

In this diff, I make the evaluation of return hint also lazy, similar to arg_list and jsx_props.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D43317488

fbshipit-source-id: e7a02e2376413f9e8b3cb57c5cd9bdf2ceaa1141