replace is_typemap_reason and is_calltype_reason checks with more robust alternatives
Summary:
In Flow_js we special-case the behavior of calls on polymorphic and intersection types based on
the reason of the callee. Reasons are not the most reliable information to use for this purpose.
Instead, this diff includes the necessary information within `Type.funcalltype`.

In a follow-up diff, we will use this information to avoid using the newly implemented LTI implicit
instantiation code for calls originating from utility types (`$Call`, `$ObjMap`, etc.).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D43382409

fbshipit-source-id: b3e6142f9e7f2d25357f1dcfca34bb8d3c20c13c