Change it so `symbol` cannot be a subtype of an interface
Summary:
`symbol` is a primtiive type (for an `x` of that type, `typeof x === 'symbol'`), I don't think it is super useful for it to be able to be a subtype of an interface (and as such a subtype of `interface {}`).

I previously numbers, booleans, etc. in D35711007 (https://github.com/facebook/flow/commit/7128b3751219cc665bda69b42e860836df46e96f)

I think it is a reasonable goal to have `interface {}` be things which are `typeof` `"object"` or `"function"` (excluding `null`).

After this change, only `string` remains as an exception to the above.

Changelog: [errors] `symbol` can no longer be a subtype of an interface

Reviewed By: panagosg7

Differential Revision: D43400881

fbshipit-source-id: f34781818cd9ab0ad2cdda99b2c14d778100d2b7