[flow-upgrade] Create codemod to convert `$Shape` to `Partial`
Summary:
Changelog: [flow-upgrade] Add codemod to convert usage of the `$Shape` utility type to `Partial`. It is possible that this will create type errors that you will have to resolve.

This won't be added to the specific "upgrades" list until we are actually going to kill `$Shape`. It is available to run with the `yarn run flow-codemod`

Reviewed By: SamChou19815

Differential Revision: D43522006

fbshipit-source-id: 5224fc3238713fe914e2edd32722e1e8effe63a2