[conditional-type] Parsing support
Summary:
This diff introduces the parsing support for conditional type. I'm using the exact same syntax as TypeScript, and the exact precedence level (lowest).

Type checking is not done: we just error and produce any is the flag is not enabled, and produce any for now when the flag is enabled (so that we can test the flag).

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D42512544

fbshipit-source-id: c6f95337e02e3c2bace1fc5bdf18c49a1c173dda