[IDE] Add local rename feature to LSP
Summary:
This diff enables the LSP "rename" feature for local renaming. Right now, it only works in a single file. It takes the most conservative approach in trying to rename the fewest things when it runs into named imports or object shorthands (in destructuring or object literals). We may eventually want to create an option to "follow" renames across shorthands but I'll wait for feedback on what developers prefer before adding this feature.

The implementation of this feature relies on the "find references" command also added recently. It takes these references, checks if any of them are shorthands and, if they are, expands the shorthand with the old and new name. Otherwise, it just replaces the reference with the new name.

Changelog: [Feature] Add local variable renaming capabilities to the LSP

Reviewed By: SamChou19815

Differential Revision: D43200269

fbshipit-source-id: 14bd7c1604c7cf43216b81453986c046a9998065