Attempt to cache more expressions when in synthesis mode
Summary: We currently cache expressions only when we're not in synthesis mode. I believe (but am not certain, please check me on this!) that it's safe to do this cacheing if the result of typechecking an expression is free of placeholders and no errors were created, even if elsewhere in the same synthesis mode placeholders were created--cf similar logic for `synthesize_expression_for_instantiation`. This diff performs this additional cacheing; to do so, we now need nested `run_in_synthesis_mode` calls, so we add an option to merge the placeholders generated by nested calls rather than restoring the original (so that if we enter a nested synthesis mode for an expression and it generates a placeholder, the overall toplevel synthesis mode knows that a placeholder was generated as well.

Reviewed By: SamChou19815

Differential Revision: D43421885

fbshipit-source-id: aa5229bfc7e30efe1a5236edf8cde485ed863c2c