[website] Enable `js flow-check` on more snippets
Summary:
As a preparation step to enable LTI for all snippets, I made a pass to enable more inline errors for more snippets, so that it's unlikely that some new LTI errors will slip through.

This diff is a wave that simply adds the check (with some cosmetic fixes). Further waves are needed to fix some pre-existing issues.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D43580757

fbshipit-source-id: 73ff4395aa1a46bed8374546ca83ce8f50e9301d