Do not consider TaggedTemplates to be definitely synthesizable
Summary:
TaggedTemplates are function calls, which are not definitely synthesizable. Returning false here makes it so that we record hints when visit TaggedTemplates:
https://github.com/facebook/flow/blob/b0af5459505640107574344cdb37b224b7799f76/src/analysis/env_builder/name_def.ml#L2380-L2384

We need to be able to thread hints in order to fix certain classes of errors, like underconstrained implicit instantiations. See the added test for an example.

Changelog: [fix] We fixed a bug preventing annotation hints from being used in tagged template expressions. Threading hints correctly may get rid of some underconstrained-implicit-instantiation errors that were otherwise impossible to fix.

Reviewed By: SamChou19815

Differential Revision: D43666330

fbshipit-source-id: 2c0aea62349606a7d1a172751f32aeb21d3c7642