[IDE] Fix rename with shorthand named imports
Summary:
This is a quick fix for a bug in the implementation of the LSP renaming action. Previously, it didn't check for whether a named import was actually making use of the shorthand syntax. It erronously expanded the identifier to be `oldName as newName` even when `oldName` was already a local name.

I also changed the name of the type to `shorthandKind` to make it clear this refers to a shorthand of either an object or an import.

Changelog: [internal]

Reviewed By: mroch

Differential Revision:
D43704868 (https://github.com/facebook/flow/commit/f6ddb95d08d0e22372c9f11c1cc3ad551befc037)

------------------------------------------------------------------------
(from 144c8b075c5d77083f456cbe48a2019b012a4959)

fbshipit-source-id: ff215adc6c2d3b331da2e59648a8841b69583ea8