[IDE] Use AST differ in rename action
Summary:
Previously, the rename LSP command was just creating AST nodes and sending these nodes to the IDE alongside the location. This is error prone as it allows for the creation of invalid ASTs. By using the mapper to update the AST then creating the changes with the AST differ, we avoid these potential bugs.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D43782969

fbshipit-source-id: ecd8bd0af0020e9dd6324c0dbcef8c9e48788343