[lti] Remove placeholder fallback for out-of-outer reads in synthesis mode
Summary:
This shouldn't happen since we ensured it in name_def_ordering by declaring dependency on all the names. We should restore this sanity check.

When enabling it, it revealed a hidden `MissingEnvWrite` error not noticed before, when we try to read the index `b` in `a[b] = expr` for hint on `expr`, but we never create a write entry for `b`. This diff also fixes this issue.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D43368777

fbshipit-source-id: 3695a19713a37923501bb3f865fd75c13afb3325