Kill Daemon.fork
Summary:
This is unused. We always spawn processes instead.

The fork and spawn code both call `setup_channels`, which sets up pipes or
socketpairs per worker. By removing this code, we can clean up the whole pipes
vs. sockets mess, explained right below:

Currently we use pipes, but this causes problems on Windows+Lwt where we try to
use Unix.select. Windows `select` only works with sockets. OCaml tries to
emulate POSIX behavior, but it's buggy (edge triggered instead of level
triggered).

There's no great option for doing async with Windows on OCaml right now. Lwt's
select-based methods are broken, but so is it's libuv backend. Eio is a
promising option for OCaml 5, but even they are figuring out the Windows story
(they still use the buggy libuv on Windows).

So, for now, it seems like the best option is to use sockets everywhere, so that
Unix.select works on Windows. This is what mroch explored in D40862093.
However, I think it's kind of gross to create two socketpairs for each worker.
Sockets are bidirectional!

Instead, I want to create a single socketpair when channel_mode is socket, but
that makes the branches between pipe and socket even greater, since now the
cloexec stuff is also different...

Reviewed By: mroch

Differential Revision: D43895867

fbshipit-source-id: 53c60efeaedfdf3913f3fe3c313572cec01acffc