Intern strings in saved state
Summary:
Saved states are not very normalized. For example, we use File_key.t for keys,
so a given file name can appear many times. Compression helps eliminate this
duplication in transit, but duplication in the decompressed value is also
problematic.

Before this diff, we interned File_key.t values specifically. This diff
generalizes and extends this to all strings.

Reviewed By: mroch

Differential Revision: D43925105

fbshipit-source-id: 4af6f8fbcfbf8d123d9cfe3e049c71dd4a5dae80