Remove global master context cache
Summary:
Before this diff, we used the SharedMem.WithCache API for the master context
heap. This API creates a module-level cache which is never cleared.

This diff replaces the cache by instead fetching the master cx once and passing
it down into the places that need it.

1. The server process uses the master cx to handle IDE requests. For this use
case, we store the master cx in the server environment.

2. For check workers, we read the master cx once per job and pass it into the
check service.

Motivation:
I want to kill the WithCache API entirely, and this is the last remaining user.
See the rest of this stack for cleanup diffs.

Also, in preparation for long-lived workers, we need to be more careful about
global caches. By being explit about where this data is fetched, we can be
confident that worker GC will collect it.

Future work:
Eventually, I expect we will eliminate the "master cx" entirely. Instead of
storing a context, we have discussed an architecture more similar to types-first
2.0, where we store the builtin type sig instead and materialize types on
demand.

If we do move to builtin signatures instead of master cx, then most of this code
would just go away, and the "check cache" would be responsible for holding on to
the materialized type information.

Reviewed By: mroch

Differential Revision: D43890542

fbshipit-source-id: 526ac531c98357ed7dba0a3f50b4f6eb83829c60