Move LocalCache out of SharedMem
Summary:
The LocalCache stuff used to be part of WithCache, which is removed earlier in
this stack. Now it is only used separately in a few places to cache specific
things: ASTs and aloc tables.

Moving this code out of sharedMem.ml just for organization purposes. This is
kind of complicated code, so moving it into a separate file makes it easier to
read (IMO) and easier to browse sharedMem.ml (IMO).

Reviewed By: mroch

Differential Revision: D43890883

fbshipit-source-id: ee633864e1b436c0aae42c1378b32540664dc278