Remove sharedmem serialize/deserialize logging and config
Summary:
This diff removes the "shared memory log level" setting and the logic controlled
by the setting.

When the log level was 1, we would measure the serialized/compressed size and
decompressed size for "serialized" heap objects. When log level was 2, we would
further traverse the deserialized objects to get their full size in OCaml.

This code only fires for the "old" heap code, where we store blobs containing
serialized OCaml values. But this code is being replaced with the "new" heap, so
this code is no longer very useful, since it only describes a fraction of shared
memory.

Instead of extending this logging to the new heap, I propose that we remove it
entirely. Our shared memory usage is very much under control these days. This
code was added when shared memory usage was a huge problem and we needed lots of
visibility into it.

The config setting that controlled this behavior was public API and documented,
but I propose we just yeet it instead of attempting to deprecate. I don't think
anyone except the Flow team ever used it.

Changelog: [breaking] Removed the "sharedmemory.log_level" flowconfig option and "--sharedmemory-log-level" command line option

Reviewed By: mroch

Differential Revision: D43893761

fbshipit-source-id: 0a9ad3304b640b18a9064951919b878f53552067