pull out "typify" of Ty.elt into utility
Summary:
In a few places around the codebase, we were converting `Ty.Decl (Ty.ClassDecl (s, _))` to
`Ty.TypeOf (Ty.TSymbol s)`, so that instead of adding `class C` as a type during codemods,
which is not valid syntax, we would insert `typeof C`.

This diff pulls out the code that does this to a Ty_utils function. It also includes the case of
`EnumDecl` (`enum E`), for which we could also use the same behavior.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D43992781

fbshipit-source-id: 13303127c8711f029eba4bb7ab2a55d6dcc6f49b