[conditional-type] Parse infer type
Summary:
This diff implements the logic to parse infer types. This is more complicated than it appears to be, because in TS 4.7 they added support for optional bounds on `infer T` (https://devblogs.microsoft.com/typescript/announcing-typescript-4-7/#extends-constraints-on-infer-type-variables).

For the parsing logic, I mainly consulted swc's implementation as follows:

https://github.com/swc-project/swc/blob/2525a844da561f5e6feee35d27f690e0b4881abc/crates/swc_ecma_parser/src/parser/typescript.rs#L923-L940
https://github.com/swc-project/swc/blob/2525a844da561f5e6feee35d27f690e0b4881abc/crates/swc_ecma_parser/src/parser/typescript.rs#L2251-L2280

I also checked the original TS implementation and verified that it's correct.

The parsing algorithm requires backtracking. After we parsed `infer some_id`, it will try to parse `extends non_conditional_type`. If it is followed by `?` (which signals the start of a conditional type), we will back track, and consider the infer type to be `infer some_id`, otherwise, it is `infer some_id extends non_conditional_type`. There is also the parser env flag `no_conditional_type` that we have to maintain. I just followed the reference implementation on when to set or reset them.

I also need to propagate `from_infer_type` information to `mk_type_param`, so that it doesn't error when the name and bound is separated by `extends`

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D43589818

fbshipit-source-id: 2da33e9372b4bce1e264df59551979eefb82a99a