Include function predicate in FunT type
Summary:
When support for predicate functions was first added, Flow used to process predicate code alongside
checking. This was because in code like
```
function is_number(x: mixed): %checks {
  return typeof x === "number";
}
```
it wasn't possible to know the predicate encoded in the return expression until we inferred a type for it.
Historically, this happened in Statement.predicates_of_condition. It was therefore convenient at that time to
incorporate the inferred predicate in the return type itself. When encountering a return expression, we'd
extract the predicates that emerged from them and wrap them in an `OpenPredT`.

With the new environment that rolled out in the earlier stages of LTI, we are in a position to infer the predicates
associated with predicate functions much earlier -- before we even start type checking.

This diff leverages this fact to eliminate the `OpenPredT` type. Instead the predicate encoded by the function
is stored in the `funtype` payload of the `FunT` constructor. This simplifies processing of function predicates
as we can eliminate the need for setting up constraints (e.g. `CallOpenPredT`) in favor of direct operations on
predicates.

In follow up diffs we'll be cleaning the constructors `OpenPredT`, `SubstOnPredT` and `CallOpenPredT`.

Changelog: [intenal]

Reviewed By: SamChou19815

Differential Revision: D44029776

fbshipit-source-id: 79944e138f09d4b38b811cbebc0d4433d863e22c