Add definition of Performance.eventCounts according to the W3C standard
Summary:
Pull Request resolved: https://github.com/facebook/flow/pull/9021

Expand the `Performance` typing with the `eventCounts`, [according to the W3C standard](https://www.w3.org/TR/event-timing/#sec-event-counts) (more detailed API description [can be found here](https://developer.mozilla.org/en-US/docs/Web/API/EventCounts)).

This is something that is implemented for web, and now we have it implemented in React Native as well, with the goal to be used in the performance APIs migrated from Comet.

Reviewed By: mroch

Differential Revision: D44059718

fbshipit-source-id: 5dfcb22ecd6fcc7f7dbf75d0334755bd60302e8b