[IDE] Add support for willRenameFiles action
Summary:
This diff sets up the server capablities to use the `workspace/willRenameFiles` action. When enabled, the client will send a request to the server to get workspace changes before applying a file rename. This will allow the server to do a rename of imports when a module name is changed.

To avoid the client sending us requests right now, I've left the list of `filters` empty. When we turn this feature on, it should be updated to handle Flow files.

Reviewed By: mroch

Differential Revision: D44119864

fbshipit-source-id: 316400b30c5cbd54bd46a4e209fdd917a9bc768e