don't restart when out of shared memory
Summary:
running out of shared memory currently restarts the server, instead of killing the monitor.

in theory, this might clean out enough stuff that the server can re-init and run for a while. in practice, this is a serious problem that shouldn't happen, and we shouldn't be silently failing and restarting.

now, the monitor will exit instead of restarting the server. `flow`/`flow start` will retry to start a server a few times, and this will consume a retry, so if we run out of memory on init, it will surface to the user instead of crash-looping.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D44233304

fbshipit-source-id: bdff1f61cf24d35e30256f2130d1b370a6b60325