Homomorphic Mapped Types type checking
Summary:
This diff adds type checking support for homomorphic mapped types with the following limitations:
1. No additional properties
2. No exactness modifiers-- we take on the exactness of the mapped object
3. No `-?` optional modifier removal. This requires a bigger refactor in how we model optional properties
4. only homomorphic mapped types using an inline keyof (keyof is not a standalone type)
5. Not supported in declared classes or interfaces

We use the normal object_kit machinery to turn the target of the `keyof` into a slice and perform a typeapp to implement the mapping function. See slice_utils for the implementation. I kept it in slice_utils to make sure I never accidentally use `Flow_js` so that mapped types will work in annotation inference.

I do not remove the "Mapped Types are not supported" error because this diff does not implement type sig support. Once that step is done I will remove that error.

**The most meaningful changes are in slice_utils.ml**. type_annotation.ml is also interesting for interactions with the typed ast.

The remaining work for homomorphic mapped types is:
1. Ty normalization
2. Type sig support
3. Implicit instantiation reversals

1 & 2 are required for the MVP and I will follow up with 3 later.

Reviewed By: SamChou19815

Differential Revision: D43988345

fbshipit-source-id: 7dc38eec41e4076fdd34979025b38ca7fe86f03b