[mapped-types] Keep track of key locations in slices for better error messages in mapped types
Summary:
Mapped Types create SingletonStrTs to substitute into the property definition. Slices did not keep track of the location of the keys, which forced error messages to be positioned at the property instead. This diff adds key_loc information to slices and uses that location to create the SingletonStrT.

I do not make an effort to ensure that this field is always set by other operations in slice-utils, but did persist the key_loc when the solution was obvious. At worst, this means that some mapped type errors involving the key type (expect relatively rare) may be poorly positioned, but no worse than they were before this diff. This diff enables us to fix those poorly positioned errors by thinking through the cases I punt on.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D44184254

fbshipit-source-id: b323a6768e5b517f2fa2d6ce871356d48dd35b43