gracefully shut down sockets on daemon exit
Summary:
when using sockets instead of pipes to communicate with daemons, we ran into an issue where reading from the daemon's socket raised `ECONNRESET` on Windows.

Winsock raises `WSAECONNRESET` when the connection is "reset by the remote side executing a hard or abortive close."

https://learn.microsoft.com/en-us/windows/win32/winsock/graceful-shutdown-linger-options-and-socket-closure-2 discusses this in some detail.

it seems that calling neither `shutdown` nor `closesocket` before exiting can lead to a hard close. so, this diff is calling `closesocket` on exit.

it also seems to work to call `Unix.shutdown (descr_of_out_channel oc) Unix.SHUTDOWN_SEND`, but that only works on sockets. `close` works on both sockets and pipes. I went with `close` to avoid having to know what kind of channel we have.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D44306081

fbshipit-source-id: 861b23efb630caecdbe0b15ff9443b53d22d659d