Quick fix from `$Partial<T>` to `Partial`
Summary:
Changelog: [ide] Adds a quick fix from the deprecated `$Partial` to `Partial`

The implementation here is a bit heavy weight, but this is because it is intended to very easily extended to additional deprecated utility type quick fixes. E.g. `$Shape` -> `Partial`, and in the future stuff like `$ReadOnly` -> `ReadOnly`/`Readonly`, etc. - we will have many of these. All you will need to do is add a variant to `Error_message.DeprecatedUtility` defined in the previous diff.

Reviewed By: panagosg7

Differential Revision: D44317891

fbshipit-source-id: 6618f1bdf52389a3d4bdc0bb55623507e5ceeb23