[IDE] Refactor find_local_refs logic
Summary:
This diff is just a refactor of the shared code used in find all refs, document highlight, and rename. It seemed strange to return the parse information along with the local ref information so split these into two functions. I then factored out the shard code in document highlight and find all refs to a mapping function. This will clean things up for working on global rename.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D44150113

fbshipit-source-id: 385c7ae2ef0bf6fd979bae111ca687a13c0c4d46