Add MappedType to AST
Summary:
This diff adds a MappedType AST node and addresses all the cases needed to make the code compile again. It does not add parsing/type sig/type checking support. I am separating this change out to keep the diffs smaller.

This diff models Mapped Types as a property to give us the option to support mapped types with additional properties in the future. We do not follow TS's ast output here, so this will necessitate changes in external tools that babel and prettier. We are making this decision deliberately because we do not want to be limited by the TS design decisions if we choose to diverge, and it's likely that our design will not match theirs 1:1.

Reviewed By: gkz

Differential Revision:
D43509418

------------------------------------------------------------------------
(from b14182b2ee6aab686f197c049b37ed0e6b8b7167)

fbshipit-source-id: a82475c0940091592c6120f32067179e240e2c44