fix on nested EvalT operations
Summary:
When `mk_type_destructor` is passed in an `OpenT` it introduces a `TypeDestructorTriggerT` (TDT)
between the OpenT and the actual use U that corresponds to the destructuring. One of the positive effects
of using TDT this way is that it acts as a barrier intercepting union-like types and preventing them from
being split up before they are forwarded to U.

EvalTs themselves were not handled as OpenTs in `mk_type_destructor` even though they evaluate
to an OpenT. This led to surprising results whenever nested EvalTs would occur. Take for example the following
code
```
type T = {a: ?{b: ?{x: string, y: number}}};
type O = $Diff<$NonMaybeType<T['a']?.['b']>, {y: mixed}>;
const obj: O = {x: 'hi'};
```
The check in the end becomes a `l ~> ReposUseT ({x: 'hi'})` constraint, where `l` will receive the result
of the nested EvalT computation in `O`. It is important to not break up any unions that emerge in `O`
before they hit the `ReposUseT`.

However, this was not guaranteed before this diff:
(a) `T['a']?.['b']` would return a Union of a MaybeT and a VoidT
(b) `$NonMaybeType<.>` would take the above union and split it up as it passed it through `FilterMaybeT`

In the end, we ended up with an OpenT with two lower bounds: an EmptyT and an ObjT. In general this is
not necessarily a problem, except we had already set up a `ReposUseT` that reverses each incoming type
consequtively. So `EmptyT ~> ReposUseT ({x: 'hi'})` would become `{x: 'hi'} ~> EmptyT`
which is an error.

What this diff achieves is change step (b) above into

(b') Flow the Union of a MaybeT and a VoidT to a TDT. This constraint is special in that it preserves the
union-ness of the input. (See `eval_destructor` and `destruct_union`.)

In the end, now we end up with a union flowing to `ReposUseT`.

Changelog: [fix] fixed a bug that caused instances of nested utility types to be considered equivalent to `empty` (for example [try-Flow](https://flow.org/try/#0JAKALgngDgpgBAFTgXjgbwIYC44H40BGO+AHjgM5gBOAlgHYDmANHBDnQK4C2BMVAvoIDcIcNHgAhFHAAkAOQD2dALIYIvBOIA8CANoByDPoC6uAHQGCJgHwiAxkspwicKajRk4+gBY19LNjgAZn4hOAB6cLg6BTg+KgUqUUhYOAB5AEZpGQARGgAzfK0JFjRArhoSGAATflsQBzonBQIAKwycTOkPHB8-UIiomLiqBKSxVLSAJmy8wq15JVV1GE1YHQMjUwt9K2NrUvLKmrqRSLgLgD1XAFUAcRwwX3I4FPgAdwwXxvIaar4anEuFBIA1HGA4C1WlNOjN3J4+voBudyLEnjQXl9fgw6FwYHQIVQMBiYJi6CMxkA))

Reviewed By: jbrown215

Differential Revision: D44326794

fbshipit-source-id: b60de2f7072bb802e193ddbf36cb35746f76212d