Always error on the existential type `*`
Summary:
The existential type `*` has been deprecated and just an alias for `any` since version 0.163. However, it was only an error to use it if you enabled the `unclear-type` Flow lint rule. If you had not enabled it, there would be no error.

Changelog: [errors] Always error on the deprecated existential type `*`, it's just an alias for `any`

Reviewed By: SamChou19815

Differential Revision: D44276187

fbshipit-source-id: 0476be1b23f42d6613e9293d24b11b7f640f2163