[try-flow] Make type checking dynamically configurable, and remember config and version in URL
Summary:
This diff sets up the foundation for a dynamically configurable try-flow. It is designed to reduce future maintenance cost.

On the ocaml side, checking related functions will take an additional argument for configuration. In addition, it will produce a schema that should be in sync with the shape of the config object. This setup allows us to easily update the list of config on the ocaml side, while keeping the website code intact across versions. The new argument will be the last argument, so that the older versions can gracefully fallback to ignoring whatever that's passed into that position.

On the js side, we will dynamically render the schema as an editor, and produce a config object that can be fed into flow.js. I have also made some significant refactoring to fix some of the convoluted code I wrote before. I also encoded the version and config into the url. I'm currently using the encoding of `{code,version,config}`, and compress the `JSON.stringify` result.

Changelog: [try-flow] Try flow's behavior can now be configured

Reviewed By: gkz

Differential Revision:
D43072921

------------------------------------------------------------------------
(from 522f1d06fa0fbc08c1c01e1109f4972fc13ba5cf)

fbshipit-source-id: 4db8046d4b5d3fee80ecb0ce35561fcb628588b4