Remove unused this_type_params ref
Summary:
This was added in D32366997 to bind the "this" tparam in function signatures. In
D38925514, we removed binding for type parameters _except_ for "this".

On a hunch, I tried just removing the "this" binding as well, and it seems to
work, but to be honest I don't understand why.

Currently, the Env.this_type_params value is a memory leak because it's never
cleared.

Reviewed By: SamChou19815

Differential Revision: D44361582

fbshipit-source-id: d85a515ea63b7cfa1d8e8a6f32cdb73bced01c8e