Fix tvar_cache leak in coverage service
Summary:
The visitor object is allocated in the module scope and re-used. The object has
local state, a tvar cache, which records information for every tvar visited.

This cache is never cleared, so the memory leaks over time. For check jobs, this
isn't too bad because we kill the process after each batch of files. However,
this leak is a big problem for long-lived workers.

Reviewed By: mroch

Differential Revision: D44354176

fbshipit-source-id: e14c8bff7c89b83aca7dff61e59dd2b917378397